found another rand error

+++ src/cpu/kernel.c
@@ -30,7 +30,7 @@ unsigned long long run_kernel( Input * I, Source * S, Table * table)
 		// Allocate Thread Local Flux Vector
 		for( int i = 0; i < I->egroups; i++ )
 		{
-			state_flux[i] = rand_r(&seed) / RAND_MAX;
+			state_flux[i] = (float) rand_r(&seed) / RAND_MAX;
 
 			#ifdef VERIFY
 			state_flux[i] = 1.0;
